Base logic for course models redefinition
